Fix bugs uncovered by compiler warnings
clang 3.5 and GCC 4.9 have been used with -Wall -Werror


git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/soma/soma-io/branches/bug_fix@69235 613412ce-1361-4b7a-93bf-a3c75af1bae4

+++ src/somaopenslide/config/osversion.cc
@@ -52,10 +52,10 @@ string soma::openslideShortVersion()
   static string	ver;
   if( ver.empty() )
     {
-    	string longver = OPENSLIDE_VERSION;
-      unsigned found = longver.rfind('.');
-		  if (found != std::string::npos)
-    		ver = longver.substr(0, found);
+      string longver = OPENSLIDE_VERSION;
+      size_t found = longver.rfind('.');
+      if (found != std::string::npos)
+        ver = longver.substr(0, found);
     }
   return ver;
 }
